feat: Orientation control
Calculate quaternion needed to make the drone tangential to the circle.
A testing python script is written to verify my calculation.

One problem now is that I have to reduce theta by half in order to get
the right result (both in python script and sitl program).

Another problem is that during sitl, the drone's quaternion sometimes
went to negative value of my command.