test: Wrap zed application as a class
The code is based on this repository:
https://github.com/adujardin/zed-save-depth/blob/master/src/main.cpp

Because we will develop more sophisticated obstacle detecting
application, I wrap this program as a class. Now that is just a
prototype.

Still, I need to move my work in depth_image.cpp into the class. After
that we will have a concrete foundation for developing ZED applications.