fix: Debug control flow of navigation node
Hard lession: thread::join() only returns when the thread completed.

The problem was Navigator::Ok(). state_.armed should not be included in
the function.

Also, Navigator should not listen to high-level channel. In the future,
I will rewrite the AI node so that the node would translate the
high-level commond to setpoints.